Add tons of edits, tweek CSS/Html for marquee & slideIn animation.

+++ app/db/insert_data.py
@@ -4,6 +4,7 @@ from sqlalchemy.orm import sessionmaker
 from Tables import Base,Edit
 import requests as rq
 import re
+from pathlib import Path
 
 file_path = os.path.abspath(os.getcwd())+"/app/db/app.db"
 engine = create_engine('sqlite:///'+file_path)
@@ -11,6 +12,7 @@ Base.metadata.create_all(engine)
 Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+counter = 0
 
 def import_from_csv(csv_filepath):
   rows = []
@@ -23,8 +25,8 @@ def import_from_csv(csv_filepath):
       insert_download(row)
 
 def insert_download(row):
-  if row[4] == '':
-    print("it's null!")
+  global counter
+  counter = counter + 1
   entry = Edit(
     url=row[0],
     creator_tag=row[1],
@@ -32,8 +34,13 @@ def insert_download(row):
     anime_name=row[3],
     song_name=row[4],
     song_artist=row[5])
-  session.add(entry)
-  session.commit()
+  try:
+    session.add(entry)
+    session.commit()
+  except Exception as e:
+    # print(str(counter),e)
+    print("error w/ db")
+    session.rollback()
   download(fetch_content(row[0]), str(entry.id))
 
 def check_text(cell):
@@ -48,14 +55,17 @@ def get_response(url):
   return r.text
 
 def get_urls(matches):
+  print(match.replace("\\u0026","&") for match in matches)
   return list({match.replace("\\u0026","&") for match in matches})
 
 def fetch_content(url):
     content = []
     response = get_response(url)
+    print(response)
     vids = re.findall('"video_url":"([^"]+)"',response) # find all video url links and filter out ""
     imgs = re.findall('"display_url":"([^"]+)"',response) # find all video url links and filter out ""
     content = [get_urls(vids), get_urls(imgs)]
+    print(vids, content[0])
     for c in content:
       if c:
         print('Found content: \n{0}'.format('\n'.join(c)))
@@ -71,14 +81,19 @@ def download(content, id):
   # convert_id_to_5_digits(id)
   img_path = os.path.abspath(os.getcwd())+"/app/static/images/"+id+".jpg"
   vid_path = os.path.abspath(os.getcwd())+"/app/static/videos/"+id+".mp4"
-  for v in content[0]:
-    r = rq.get(v)
-    with open(vid_path+"", 'wb') as f:
-      f.write(r.content)
-  for i in content[1]:
-    r = rq.get(i)
-    with open(img_path+"", 'wb') as f:
-      f.write(r.content)
+  print("Looking for " +  vid_path + '\n' +  "Looking for " + img_path)
+  if not Path(vid_path).is_file():
+    print(vid_path + " does not exist!")
+    for v in content[0]:
+      r = rq.get(v)
+      with open(vid_path+"", 'wb') as f:
+        f.write(r.content)
+  if not Path(img_path).is_file():
+    print(img_path + " does not exist!")
+    for i in content[1]:
+      r = rq.get(i)
+      with open(img_path+"", 'wb') as f:
+        f.write(r.content)
 
 # import from csv
 if len(sys.argv)-1 == 1: # 1 paramter
@@ -105,5 +120,5 @@ else:
     if check.lower() == "y":
       valid_input = True
   insert_download(params)
-  print("Inserted successfully!")
+  print("Script complete.")
